not ok

+++ src/Controller/CommentController.php
@@ -10,45 +10,45 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
 final class CommentController extends AbstractController
 {
-    // Affiche la liste des commentaires
     #[Route('/comment', name: 'app_comment')]
     public function index(CommentRepository $repo): Response
     {
-        $comments = $repo->findAllOrdered();
+        $comments = $repo->findAll();
 
         return $this->render('comment/index.html.twig', [
             'comments' => $comments,
+            'app_user' => $this->getUser(),
         ]);
     }
 
-    // Création d'un nouveau commentaire
     #[Route('/comment/new', name: 'comment_new')]
     public function new(Request $request, EntityManagerInterface $em): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY'); // uniquement pour utilisateurs connectés
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY'); // utilisateur connecté
 
         $comment = new Comment();
+        $comment->setAuthor($this->getUser()->getEmail());
+
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($comment);
             $em->flush();
-
             $this->addFlash('success', 'Commentaire ajouté !');
+
             return $this->redirectToRoute('app_comment');
         }
 
         return $this->render('comment/new.html.twig', [
             'form' => $form->createView(),
+            'app_user' => $this->getUser(),
         ]);
     }
 
-    // Édition d'un commentaire
     #[Route('/comment/{id}/edit', name: 'comment_edit')]
     public function edit(Comment $comment, Request $request, EntityManagerInterface $em): Response
     {
@@ -60,49 +60,27 @@ final class CommentController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
             $this->addFlash('success', 'Commentaire modifié !');
+
             return $this->redirectToRoute('app_comment');
         }
 
         return $this->render('comment/edit.html.twig', [
             'form' => $form->createView(),
-            'comment' => $comment
+            'app_user' => $this->getUser(),
         ]);
     }
 
-    // Suppression d'un commentaire
     #[Route('/comment/{id}/delete', name: 'comment_delete', methods: ['POST'])]
-    public function delete(Comment $comment, EntityManagerInterface $em): Response
+    public function delete(Comment $comment, Request $request, EntityManagerInterface $em): Response
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
 
-        $em->remove($comment);
-        $em->flush();
-        $this->addFlash('success', 'Commentaire supprimé !');
-
-        return $this->redirectToRoute('app_comment');
-    }
-
-    // Page login
-    #[Route(path: '/login', name: 'app_login')]
-    public function login(AuthenticationUtils $authenticationUtils): Response
-    {
-        if ($this->getUser()) {
-            return $this->redirectToRoute('app_comment'); // redirige si déjà connecté
+        if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
+            $em->remove($comment);
+            $em->flush();
+            $this->addFlash('success', 'Commentaire supprimé !');
         }
 
-        $error = $authenticationUtils->getLastAuthenticationError();
-        $lastUsername = $authenticationUtils->getLastUsername();
-
-        return $this->render('security/login.html.twig', [
-            'last_username' => $lastUsername,
-            'error' => $error
-        ]);
-    }
-
-    // Logout
-    #[Route(path: '/logout', name: 'app_logout')]
-    public function logout(): void
-    {
-        throw new \LogicException('Cette méthode sera interceptée automatiquement par Symfony.');
+        return $this->redirectToRoute('app_comment');
     }
 }
